Move CDataStream and CAutoFile from serialize to streams
